refactor: utils 파일을 customRender로 이름 변경 #1114
